Rename calories package Try directory creation
